requirements.txt
numpy>=1.20.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
torch>=1.9.0
requests>=2.31.0
python-dotenv>=1.0.0
transformers>=4.30.0 